Add files via upload
You have been provided with a Java class (Point.java). A Point object stores the (x, y)
coordinates of a position in two-dimensional space. The class has two constructors and a
number of accessor methods to manipulate the instance variables. Your task is to write a
new class called Line that represents a line segment between two Points as depicted in
the UML diagram below:
Your Line objects should have the following methods:
• Line(Point p1, Point p2) - Constructs a new Line that contains the given two
Points
• Point getP1() - Returns this Line’s first endpoint
• Point getP2() - Returns this Line’s second endpoint
• Line(int x1, int y1, int x2, int y2) - Add a new constructor to your
class which constructs a new Line that contains the coordinates passed to the method
Point
-x: int
-y: int
+Point()
+Point(x: int, y: int)
+getX(): int
+getY(): int
Line
-p1: Point
-p2: Point
+Line(p1: Point, p2: Point)
+Line(x1: int, y1: int, x2: int, y2: int)
+getP1(): Point
+getP2(): Point
+getSlope(): double
• double getSlope() - Returns the slope of this Line. The slope of a line between
points (x1, y1) and (x2, y2) is equal to (y2−y1)/(x2−x1). If x2 equals x1 the
denominator is zero and the slope is undefined, so you must throw an exception in
this case
Implement a test class (LineTest.java) that demonstrates the correct functionality of each of
the methods listed above and makes use of the Point class functionality. LineTest creates
a new Line and makes use of the methods getP1() and getP2() to print the Line’s end
points (x1, y1) and (x2, y2). Use the second constructor with four parameters to create
another Line and print the line’s end points. Finally obtain the slope of the Line using the
getSlope() method.